+++ float/tests/div.rs
@@ -165,7 +165,7 @@ fn test_div_by_0() {
 #[test]
 #[should_panic]
 fn test_div_by_unlimited_precision() {
-    let _ = dbig!(1).with_precision(0).value() / dbig!(3).with_precision(0).value();
+    let _ = dbig!(1).with_precision(0).unwrap() / dbig!(3).with_precision(0).unwrap();
 }
 
 #[test]