+++ float/tests/mul.rs
@@ -132,12 +132,12 @@ fn test_mul_by_inf() {
 
 #[test]
 fn test_mul_zero_precision() {
-    let a = fbig!(0xff).with_precision(0).value();
+    let a = fbig!(0xff).with_precision(0).unwrap();
     let b = fbig!(-0xff);
     test_mul(&a, &a, &fbig!(0xfe01));
     test_mul(&a, &b, &fbig!(-0xfep8));
 
-    let a = dbig!(99).with_precision(0).value();
+    let a = dbig!(99).with_precision(0).unwrap();
     let b = dbig!(-99);
     test_mul(&a, &a, &dbig!(9801));
     test_mul(&a, &b, &dbig!(-98e2));